fix(containers): mount state volume at agent CLI dirs and extract shared Result
- Replace tmpfs mounts for ~/.claude, ~/.codex, ~/.gemini, ~/.cursor-agent,
  and ~/.cache with state volume subpath binds so agent CLI state persists
  across container restarts
- Extract duplicate Containers::Result class into shared
  app/services/containers/result.rb, removing copies from ProvisionForChat
  and ChatSessionManager
