+++ app/services/containers/chat_session_manager.rb
@@ -190,35 +190,5 @@ module Containers
         **metadata
       )
     end
-
-    class Result
-      attr_reader :data, :error
-
-      def initialize(success:, data: {}, error: nil)
-        @success = success
-        @data = data
-        @error = error
-      end
-
-      def success?
-        @success
-      end
-
-      def failure?
-        !@success
-      end
-
-      def [](key)
-        data[key]
-      end
-
-      def self.success(**data)
-        new(success: true, data: data)
-      end
-
-      def self.failure(error:, **data)
-        new(success: false, data: data, error: error)
-      end
-    end
   end
 end
