+++ app/services/containers/result.rb
@@ -1,0 +1,36 @@
+# frozen_string_literal: true
+
+module Containers
+  # Simple result object for container service method returns.
+  #
+  # Shared by ProvisionForChat, ChatSessionManager, and Provision.
+  class Result
+    attr_reader :data, :error
+
+    def initialize(success:, data: {}, error: nil)
+      @success = success
+      @data = data
+      @error = error
+    end
+
+    def success?
+      @success
+    end
+
+    def failure?
+      !@success
+    end
+
+    def [](key)
+      data[key]
+    end
+
+    def self.success(**data)
+      new(success: true, data: data)
+    end
+
+    def self.failure(error:, **data)
+      new(success: false, data: data, error: error)
+    end
+  end
+end
